will make msg

+++ lab8/include/defines.h
@@ -75,8 +75,8 @@ enum Color{
 };
 
 __attribute__((regparm(3)))
-void memcpy(void *dest,const void *src,int size){
-	for (int i = 0;i < size;++i){
+void memcpy(void *dest,const void *src,uint16_t size){
+	for (uint16_t i = 0;i < size;++i){
 		*(((char*)dest)+i) = *(((char*)src)+i);
 	}
 }
